fix: Ninja Form Renders on Video Editor Page Without Selecting a Form (#854)
* refactor: moved the video editor form layer handler to separate class

* fix: render layer only when ninja form is selected.
